회원 등록, 디테일 폼 업로

+++ ejyoo_springProject/src/main/java/com/spring/controller/MemberController.java
@@ -1,16 +1,21 @@
 package com.spring.controller;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 import java.sql.SQLException;
 import java.util.Map;
-import java.util.UUID;
 
 import javax.annotation.Resource;
+import javax.servlet.http.HttpSession;
 
+import org.apache.commons.io.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -18,8 +23,12 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.ModelAndView;
 
+import com.spring.command.MemberModifyCommand;
+import com.spring.command.MemberRegistCommand;
 import com.spring.command.SearchCriteria;
+import com.spring.dto.MemberVO;
 import com.spring.service.MemberService;
+import com.spring.util.MakeFileName;
 
 @Controller
 @RequestMapping("/member")
@@ -67,7 +76,7 @@ public class MemberController {
 		} else { 
 			status = HttpStatus.OK;
 		}
-		
+		System.out.println("result : " + result);
 		entity = new ResponseEntity<String>(result, status);
 		
 		return entity;
@@ -80,7 +89,7 @@ public class MemberController {
 		if(!(multi == null || multi.isEmpty() || multi.getSize() > 1024 * 1024 * 5)) {
 			/* 파일 저장폴더 설정 */
 			String uploadPath = picturePath;
-			fileName = UUID.randomUUID().toString().replace("-","") + ".jpg";
+			fileName = MakeFileName.toUUIDFileName(multi.getOriginalFilename(),"$$");
 			File storeFile = new File(uploadPath, fileName);
 			
 			storeFile.mkdirs();
@@ -88,7 +97,7 @@ public class MemberController {
 			// local HDD 에 저장.
 			multi.transferTo(storeFile);
 			
-			if(!oldPicture.isEmpty()) {
+			if(oldPicture != null && !oldPicture.isEmpty()) {
 				File oldFile = new File(uploadPath, oldPicture);
 				if(oldFile.exists()) {
 					oldFile.delete();
@@ -96,5 +105,83 @@ public class MemberController {
 			}
 		}
 		return fileName;
+	}
+	
+	@RequestMapping(value = "/getPicture", produces = "text/plain;charset=utf-8")
+	@ResponseBody
+	public ResponseEntity<byte[]> getPicture(String picture) throws Exception {
+		InputStream in = null;
+		ResponseEntity<byte[]> entity = null;
+		String imgPath = this.picturePath;
+		try {
+			in = new FileInputStream(new File(imgPath, picture));
+			
+			entity = new ResponseEntity<byte[]>(IOUtils.toByteArray(in),HttpStatus.CREATED);
+		} catch (IOException e) {
+			e.printStackTrace();
+			entity = new ResponseEntity<byte[]>(HttpStatus.INTERNAL_SERVER_ERROR);
+		} finally {
+			in.close();
+		}
+		return entity;
+	}
+	
+	@RequestMapping("/idCheck")
+	@ResponseBody
+	public ResponseEntity<String> idCheck(String id) throws Exception {
+		ResponseEntity<String> entity = null;
+		
+		try {
+			MemberVO member = memberService.getMember(id);
+			
+			if (member != null) {
+				entity = new ResponseEntity<String>("duplicated", HttpStatus.OK);
+			} else {
+				entity = new ResponseEntity<String>("", HttpStatus.OK);
+			}
+		} catch(SQLException e) {
+			entity = new ResponseEntity<String>(HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		return entity;
+	}
+	
+	@RequestMapping(value="/regist",method=RequestMethod.POST)
+	public String regist(MemberRegistCommand memberReq) throws SQLException, IOException {
+		String url = "member/regist_success";
+		
+		MemberVO member = memberReq.toMemberVO();
+		memberService.regist(member);
+		
+		return url;
+	}
+	
+	@RequestMapping(value="/detail",method=RequestMethod.GET)
+	public String detail(String id, Model model) throws SQLException{
+		String url = "member/detail";
+		
+		MemberVO member = memberService.getMember(id);
+		model.addAttribute("member", member);
+		
+		return url;
+	}
+	
+	@RequestMapping(value="/modify",method=RequestMethod.POST)
+	public String modify(MemberModifyCommand modifyReq, HttpSession session, Model model) throws Exception{
+		String url = "member/modify_success";
+		
+		MemberVO member = modifyReq.toParseMember();
+		
+		// 신규파일 변경 및 기존파일 삭제 
+		String fileName = savePicture(modifyReq.getOldPicture(), modifyReq.getPicture());
+		member.setPicture(fileName);
+		
+		// 파일 변경 없을 시 기존 파일명 유지
+		if(modifyReq.getPicture().isEmpty()) {
+			member.setPicture(modifyReq.getOldPicture());
+		}
+		
+		// DB 내용 수정 
+		
+		return url;
 	}
 }
